apply namespace to queries
